test(flag/user module): turn FlagModule and UserModule into DynamicModules, split e2e tests
