+++ back/flag-service/.env.example
@@ -1,3 +1,6 @@
 DATABASE_URI='mongodb://127.0.0.1:27017'
 CHANGE_COOLDOWN=5
 JWT_SECRET=changeme
+
+#AUTH_BACKEND=internal
+AUTH_BACKEND=fouloscopie
