+++ back/flag-service/src/user/AuthBackend.ts
@@ -1,0 +1,4 @@
+export enum AuthBackend {
+  FOULOSCOPIE = 'fouloscopie',
+  INTERNAL = 'internal',
+}
