+++ back/flag-service/test/util/getAvailableAppPort.ts
@@ -1,0 +1,5 @@
+let port = 3001;
+
+export function getAvailableAppPort(): number {
+  return port++;
+}
